+++ src/pages/TasksPage/components/TasksTable/components/CompleteTodoButton/Comp[eteTodoButton.tsx
@@ -1,26 +1,28 @@
 import Button from "antd/es/button";
 
-import type { Task } from "../../../../types";
-import { useCompleteTodo } from './hooks'
+import type { Task } from "pages/TasksPage/shared/types";
+import { useCompleteTodo } from "./hooks";
 
 type CompleteTodoButtonProps = {
-  taskId: Task['id'];
+  taskId: Task["id"];
   disabled: boolean;
-}
+};
 
-const CompleteTodoButton: React.FC<CompleteTodoButtonProps> = ({ taskId, disabled }) => {
-  const [completeTodo] = useCompleteTodo()
+const CompleteTodoButton: React.FC<CompleteTodoButtonProps> = ({
+  taskId,
+  disabled,
+}) => {
+  const [completeTodo] = useCompleteTodo();
 
-  return <Button
-    type="primary"
-    onClick={() => completeTodo(taskId)}
-    disabled={disabled}
-  >
-    Complete
-  </Button>
-
-
-}
-
+  return (
+    <Button
+      type="primary"
+      onClick={() => completeTodo(taskId)}
+      disabled={disabled}
+    >
+      Complete
+    </Button>
+  );
+};
 
 export default CompleteTodoButton;
